src/requirements.txt
tqdm
numpy
multiprocessing
soundfile
librosa
python_speech_features
torchi>=1.5.0
scipy
scikit-learn
webrtcvad
matplotlib
torchaudio>=0.5.0
